+++ packages/js/api-core-tests/data/index.js
@@ -1,6 +1,16 @@
 const { order, getOrderExample } = require( './order' );
 const { coupon } = require( './coupon' );
 const { refund } = require( './refund' );
+const { getTaxRateExamples } = require( './tax-rate' );
+const { getVariationExample } = require( './variation' );
+const {
+	simpleProduct,
+	variableProduct,
+	variation,
+	virtualProduct,
+	groupedProduct,
+	externalProduct,
+} = require( './products-crud' );
 const { getShippingZoneExample } = require( './shipping-zone' );
 const { getShippingMethodExample } = require( './shipping-method' );
 const shared = require( './shared' );
@@ -11,6 +21,14 @@ module.exports = {
 	coupon,
 	shared,
 	refund,
+	getTaxRateExamples,
+	getVariationExample,
+	simpleProduct,
+	variableProduct,
+	variation,
+	virtualProduct,
+	groupedProduct,
+	externalProduct,
 	getShippingZoneExample,
 	getShippingMethodExample,
 };