+++ packages/js/api-core-tests/data/refund.js
@@ -14,5 +14,5 @@ const refund = {
 };
 
 module.exports = {
-	refund: refund,
+	refund,
 };