+++ packages/js/api/src/models/products/abstract/common.ts
@@ -1,4 +1,5 @@
 import { AbstractProductData } from './data';
+import { ModelID } from '../../model';
 import {
 	CatalogVisibility,
 	ProductTerm,