+++ packages/js/api/src/models/products/index.ts
@@ -1,5 +1,0 @@
-export * from './abstract';
-export * from './shared';
-export * from './simple-product';
-export * from './variation';
-export * from './variable-product';