+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-order-apply-coupon.test.js
@@ -28,7 +28,8 @@ const runOrderApplyCouponTest = () => {
 			couponCode = await createCoupon();
 			orderId = await createOrder( { productId, status: 'pending' } );
 
-			// We need to remove any listeners on the `dialog` event otherwise we can't catch the dialog below
+			await merchant.login();
+			await merchant.goToOrder( orderId );
 			await page.removeAllListeners('dialog');
 
 			// Make sure the simple product price is greater than the coupon amount