+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-order-searching.test.js
@@ -1,5 +1,3 @@
-/* eslint-disable jest/no-export, jest/no-disabled-tests, */
-
 /**
  * Internal dependencies
  */
@@ -8,6 +6,7 @@ const {
 	searchForOrder,
 	createSimpleProduct,
 	factories,
+	createOrder,
 } = require( '@woocommerce/e2e-utils' );
 
 const searchString = 'John Doe';
@@ -93,9 +92,7 @@ const runOrderSearchingTest = () => {
 	describe('WooCommerce Orders > Search orders', () => {
 		let productId;
 		let orderId;
-		beforeAll( async () => {
-			await createSimpleProduct(itemName);
-			await updateCustomerBilling();
+		let customerId;
 
 		beforeAll( async () => {
 			productId = await createSimpleProduct('Wanted Product');
@@ -109,23 +106,6 @@ const runOrderSearchingTest = () => {
 
 			// Login and open All Orders view
 			await merchant.login();
-			await merchant.openNewOrder();
-			await page.waitForSelector('#order_status');
-			await page.click('#customer_user');
-			await page.click('span.select2-search > input.select2-search__field');
-			await page.type('span.select2-search > input.select2-search__field', 'Jane Smith');
-			await page.waitFor(2000); // to avoid flakyness
-			await page.keyboard.press('Enter');
-
-			// Get the post id
-			const variablePostId = await page.$('#post_ID');
-			orderId = (await(await variablePostId.getProperty('value')).jsonValue());
-
-			// Save new order and add desired product to order
-			await clickUpdateOrder('Order updated.', true);
-			await addProductToOrder(orderId, itemName);
-
-			// Open All Orders view
 			await merchant.openAllOrdersView();
 		});