+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-product-new.test.js
@@ -144,11 +144,8 @@ const runAddSimpleProductTest = () => {
 
 const runAddVariableProductTest = () => {
 	describe('Add New Variable Product Page', () => {
-		beforeAll(async () => {
+		it('can create product with variations', async () => {
 			await merchant.login();
-		});
-
-		it('can create variable product', async () => {
 			await openNewProductAndVerify();
 
 			// Set product data