+++ plugins/woocommerce/tests/e2e/config/jest.setup.js
@@ -1,5 +1,4 @@
-const {
-	visitAdminPage,
+import {
 	clearLocalStorage,
 	setBrowserViewport,
 	withRestApi,
@@ -35,72 +34,6 @@ async function trashExistingPosts() {
 	// Delete each post
 	for (const post of posts) {
 		await client.delete(`${wpPostsEndpoint}/${post.id}`);
-	}
-}
-
-/**
- * Uses the WordPress API to update the Ready page's status.
- *
- * @param {string} status | Status to update the page to. One of: publish, future, draft, pending, private
- */
-async function updateReadyPageStatus( status ) {
-	const apiUrl = config.get('url');
-	const wpPagesEndpoint = '/wp/v2/pages';
-	const adminUsername = config.get('users.admin.username');
-	const adminPassword = config.get('users.admin.password');
-	const client = HTTPClientFactory.build(apiUrl)
-		.withBasicAuth(adminUsername, adminPassword)
-		.create();
-
-	// As the default status filter in the API is `publish`, we need to
-	// filter based on the supplied status otherwise no results are returned.
-	let statusFilter = 'publish';
-	if ( 'publish' === status ) {
-		// The page will be in a draft state, so we need to filter on that status
-		statusFilter = 'draft';
-	}
-	const getPostsResponse = await client.get(`${wpPagesEndpoint}?search=ready&status=${statusFilter}`);
-	const pageId = getPostsResponse.data[0].id;
-
-	// Update the page to the new status
-	await client.post(`${wpPagesEndpoint}/${pageId}`, { 'status': status });
-}
-
-/**
- * Use api package to delete products.
- *
- * @return {Promise} Promise resolving once products have been trashed.
- */
-async function trashExistingProducts() {
-	await merchant.login();
-	// Visit `/wp-admin/edit.php?post_type=product` so we can see a list of products and delete them.
-	await visitAdminPage( 'edit.php', 'post_type=product' );
-
-	products = await repository.list();
-	while ( products.length > 0 ) {
-		for( let p = 0; p < products.length; p++ ) {
-			await repository.delete( products[ p ].id );
-		}
-		products = await repository.list();
-	}
-}
-
-/**
- * Use api package to delete coupons.
- *
- * @return {Promise} Promise resolving once coupons have been trashed.
- */
-async function deleteAllCoupons() {
-	const repository = Coupon.restRepository( factories.api.withDefaultPermalinks );
-	let coupons;
-
-	coupons = await repository.list();
-
-	while ( coupons.length > 0 ) {
-		for (let c = 0; c < coupons.length; c++ ) {
-			await repository.delete( coupons[ c ].id );
-		}
-		coupons = await repository.list();
 	}
 }